+++ backend/packages/prompt-fragments/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @cat-factory/prompt-fragments
 
+## 1.0.72
+
+### Patch Changes
+
+- Updated dependencies [0e1e0fa]
+  - @cat-factory/contracts@0.308.1
+  - @cat-factory/kernel@0.298.1
+
+## 1.0.71
+
+### Patch Changes
+
+- Updated dependencies [7312e0a]
+  - @cat-factory/kernel@0.298.0
+  - @cat-factory/contracts@0.308.0
+
 ## 1.0.70
 
 ### Patch Changes