+++ backend/packages/provider-s3/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cat-factory/provider-s3
 
+## 0.2.389
+
+### Patch Changes
+
+- Updated dependencies [0e1e0fa]
+  - @cat-factory/kernel@0.298.1
+
+## 0.2.388
+
+### Patch Changes
+
+- Updated dependencies [7312e0a]
+  - @cat-factory/kernel@0.298.0
+
 ## 0.2.387
 
 ### Patch Changes