+++ backend/packages/sandbox/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @cat-factory/sandbox
 
+## 0.11.149
+
+### Patch Changes
+
+- Updated dependencies [0e1e0fa]
+  - @cat-factory/contracts@0.308.1
+  - @cat-factory/agents@0.129.1
+  - @cat-factory/kernel@0.298.1
+  - @cat-factory/sandbox-fixtures@0.7.350
+
+## 0.11.148
+
+### Patch Changes
+
+- Updated dependencies [7312e0a]
+  - @cat-factory/kernel@0.298.0
+  - @cat-factory/contracts@0.308.0
+  - @cat-factory/agents@0.129.0
+  - @cat-factory/sandbox-fixtures@0.7.349
+
 ## 0.11.147
 
 ### Patch Changes