+++ frontend/app/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cat-factory/app
 
+## 0.270.2
+
+### Patch Changes
+
+- Updated dependencies [0e1e0fa]
+  - @cat-factory/contracts@0.308.1
+
+## 0.270.1
+
+### Patch Changes
+
+- Updated dependencies [7312e0a]
+  - @cat-factory/contracts@0.308.0
+
 ## 0.270.0
 
 ### Minor Changes